Split generic parameter code into pass
