+++ Il2CppInterop.Generator/Contexts/MethodRewriteContext.cs
@@ -89,9 +89,9 @@ public class MethodRewriteContext
             }
         }
 
-        if (!Pass15GenerateMemberContexts.HasObfuscatedMethods && !passthroughNames &&
+        if (!Pass16GenerateMemberContexts.HasObfuscatedMethods && !passthroughNames &&
             originalMethod.Name.IsObfuscated(declaringType.AssemblyContext.GlobalContext.Options))
-            Pass15GenerateMemberContexts.HasObfuscatedMethods = true;
+            Pass16GenerateMemberContexts.HasObfuscatedMethods = true;
 
         FileOffset = originalMethod.ExtractOffset();
         // Workaround for garbage file offsets passed by Cpp2IL
@@ -281,8 +281,8 @@ public class MethodRewriteContext
         }
 
         var address = Rva;
-        if (address != 0 && Pass15GenerateMemberContexts.HasObfuscatedMethods &&
-            !Pass16ScanMethodRefs.NonDeadMethods.Contains(address)) builder.Append("_PDM");
+        if (address != 0 && Pass16GenerateMemberContexts.HasObfuscatedMethods &&
+            !Pass17ScanMethodRefs.NonDeadMethods.Contains(address)) builder.Append("_PDM");
 
         return builder.ToString();
     }
@@ -337,13 +337,13 @@ public class MethodRewriteContext
             if (a[i].ParameterType.FullName != b[i].ParameterType.FullName)
                 return false;
 
-        if (Pass15GenerateMemberContexts.HasObfuscatedMethods)
+        if (Pass16GenerateMemberContexts.HasObfuscatedMethods)
         {
             var addressA = otherRewriteContext.Rva;
             var addressB = Rva;
             if (addressA != 0 && addressB != 0)
-                if (Pass16ScanMethodRefs.NonDeadMethods.Contains(addressA) !=
-                    Pass16ScanMethodRefs.NonDeadMethods.Contains(addressB))
+                if (Pass17ScanMethodRefs.NonDeadMethods.Contains(addressA) !=
+                    Pass17ScanMethodRefs.NonDeadMethods.Contains(addressB))
                     return false;
         }
