+++ Il2CppInterop.Generator/Passes/Pass13CreateGenericNonBlittableTypes.cs
@@ -3,7 +3,7 @@ using Mono.Cecil;
 
 namespace Il2CppInterop.Generator.Passes;
 
-public static class Pass12CreateGenericNonBlittableTypes
+public static class Pass13CreateGenericNonBlittableTypes
 {
     public static void DoPass(RewriteGlobalContext context)
     {
