+++ Il2CppInterop.Generator/Runners/InteropAssemblyGenerator.cs
@@ -67,34 +67,39 @@ internal class InteropAssemblyGeneratorRunner : IRunner
             Pass10CreateTypedefs.DoPass(rewriteContext);
         }
 
+        using (new TimingCookie("Computing generic parameter usage"))
+        {
+            Pass11ComputeGenericParameterSpecifics.DoPass(rewriteContext);
+        }
+
         using (new TimingCookie("Computing struct blittability"))
         {
-            Pass11ComputeTypeSpecifics.DoPass(rewriteContext);
+            Pass12ComputeTypeSpecifics.DoPass(rewriteContext);
         }
 
         using (new TimingCookie("Creating unboxed struct types"))
         {
-            Pass12CreateGenericNonBlittableTypes.DoPass(rewriteContext);
+            Pass13CreateGenericNonBlittableTypes.DoPass(rewriteContext);
         }
 
         using (new TimingCookie("Filling typedefs"))
         {
-            Pass13FillTypedefs.DoPass(rewriteContext);
+            Pass14FillTypedefs.DoPass(rewriteContext);
         }
 
         using (new TimingCookie("Filling generic constraints"))
         {
-            Pass14FillGenericConstraints.DoPass(rewriteContext);
+            Pass15FillGenericConstraints.DoPass(rewriteContext);
         }
 
         using (new TimingCookie("Creating members"))
         {
-            Pass15GenerateMemberContexts.DoPass(rewriteContext);
+            Pass16GenerateMemberContexts.DoPass(rewriteContext);
         }
 
         using (new TimingCookie("Scanning method cross-references"))
         {
-            Pass16ScanMethodRefs.DoPass(rewriteContext, options);
+            Pass17ScanMethodRefs.DoPass(rewriteContext, options);
         }
 
         using (new TimingCookie("Finalizing method declarations"))
